try fix logical_xid_subxacts_test checks 2

+++ sql/orioledb--1.5--1.6_dev.sql
@@ -5,12 +5,12 @@
 
 -- Get current logical xid to remember it
 CREATE FUNCTION orioledb_get_current_logical_xid()
-RETURNS int
+RETURNS bigint
 AS 'MODULE_PATHNAME'
 VOLATILE LANGUAGE C;
 
 -- Get current heap xid to remember it
 CREATE FUNCTION orioledb_get_current_heap_xid()
-RETURNS int
+RETURNS bigint
 AS 'MODULE_PATHNAME'
 VOLATILE LANGUAGE C;
